move distributions to one place
