+++ foundry/survival/__init__.py
@@ -1,2 +1,1 @@
-from .distributions import CeilingWeibull
 from .util import make_discrete_target
